+++ Pool.h
@@ -34,7 +34,7 @@ public:
 	virtual void propogate();
 	virtual void updateS();
 	virtual void updateV();
-	virtual float* getStateLocation(int, State);
+	virtual double* getStateLocation(int, State);
 	void print();
 	void toFile(string appendString);
