+++ PoolRecEx.h
@@ -12,24 +12,24 @@ class PoolRecEx: public PoolRec
 	
   public:
 	
-//	static const float aPre = 4;	
+//	static const double aPre = 4;	
 	
-	static const float tau_AMPA = 2;
-	static const float tau_AMPA_rise = 2;
-	static const float tau_NMDA = 100;
-	static float tau_AMPA_Inv_times_dt;
-	static float tau_AMPA_rise_Inv_times_dt;
-	static float one_minus_tau_NMDA_Inv_times_dt;
-	static float alpha_times_dt;
+	static const double tau_AMPA = 2;
+	static const double tau_AMPA_rise = 2;
+	static const double tau_NMDA = 100;
+	static double tau_AMPA_Inv_times_dt;
+	static double tau_AMPA_rise_Inv_times_dt;
+	static double one_minus_tau_NMDA_Inv_times_dt;
+	static double alpha_times_dt;
 	
-	static const float gext_AMPA_E = .000000001*2.1;
-	static const float grec_AMPA_E = .000000001*.05;
-	static const float gNMDA_E = .000000001*.165;
-	static const float gGABA_E = .000000001*1.3;
-	static const float gl_E = .000000001*25;
-	static const float cm_E = .000001*.5;
-	static float dt_times_gL_E_over_cm_E;
-	static float dt_over_cm_E;
+	static const double gext_AMPA_E = .000000001*2.1;
+	static const double grec_AMPA_E = .000000001*.05;
+	static const double gNMDA_E = .000000001*.165;
+	static const double gGABA_E = .000000001*1.3;
+	static const double gl_E = .000000001*25;
+	static const double cm_E = .000001*.5;
+	static double dt_times_gL_E_over_cm_E;
+	static double dt_over_cm_E;
 	
 	int i;
 	
@@ -40,11 +40,11 @@ class PoolRecEx: public PoolRec
 	~PoolRecEx();
 
 	// Member data:
-	valarray<float> *AMPA;
-	valarray<float> *NMDA;
-	valarray<float> *X;
-	float AMPA_pooled;
-	float NMDA_pooled;
+	valarray<double> *AMPA;
+	valarray<double> *NMDA;
+	valarray<double> *X;
+	double AMPA_pooled;
+	double NMDA_pooled;
 	
 	// Member functions:
 	void propogate();
