+++ PoolRecInh.h
@@ -12,16 +12,16 @@ class PoolRecInh: public PoolRec
 	
   public:
 	
-	static const float tau_GABA = 5;
-	static float tau_GABA_Inv_times_dt;
-	static const float gext_AMPA_I = .000000001*1.62;
-	static const float grec_AMPA_I = .000000001*.04;
-	static const float gNMDA_I = .000000001*.13;
-	static const float gGABA_I = .000000001*1.0;
-	static const float gl_I = .000000001*20;
-	static const float cm_I = .000001*.2;
-	static float dt_times_gL_I_over_cm_I;
-	static float dt_over_cm_I;
+	static const double tau_GABA = 5;
+	static double tau_GABA_Inv_times_dt;
+	static const double gext_AMPA_I = .000000001*1.62;
+	static const double grec_AMPA_I = .000000001*.04;
+	static const double gNMDA_I = .000000001*.13;
+	static const double gGABA_I = .000000001*1.0;
+	static const double gl_I = .000000001*20;
+	static const double cm_I = .000001*.2;
+	static double dt_times_gL_I_over_cm_I;
+	static double dt_over_cm_I;
 	
 	int i;
 	
@@ -32,8 +32,8 @@ class PoolRecInh: public PoolRec
 	~PoolRecInh();
 
 	// Member data:
-	valarray<float> *GABA;
-	float GABA_pooled;
+	valarray<double> *GABA;
+	double GABA_pooled;
 	
 	// Member functions:
 	void propogate();
